Segfault fixed pour just 1 philo case

+++ include/struct.h
@@ -6,7 +6,7 @@
 /*   By: mlaffita <marvin@42lausanne.ch>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/15 13:59:39 by mlaffita          #+#    #+#             */
-/*   Updated: 2025/07/29 16:42:30 by mlaffita         ###   ########.fr       */
+/*   Updated: 2025/08/04 14:25:14 by mlaffita         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ typedef struct s_philo
 {
 	int				philo_id;
 	int				meal_num;
-	int				last_meal;
+	long			last_meal;
 	pthread_t		thread;
 	pthread_mutex_t	*left_fork;
 	pthread_mutex_t	*right_fork;
